+++ src/programs/ensembleH.f90
@@ -36,20 +36,31 @@ program midas_ensembleH
   use obsErrors_mod
   use innovation_mod
   use ensembleObservations_mod
+  use ensembleStateVector_mod
+  use enkf_mod
   implicit none
 
-  type(struct_obs), target :: obsSpaceData
-  type(struct_gsv)         :: stateVector, statevector_tiles
-  type(struct_columnData)  :: column
-  type(struct_eob)         :: ensObs, ensObs_mpiglobal
+  type(struct_obs), target  :: obsSpaceData
+  type(struct_ens)          :: ensembleTrl4D
+  type(struct_gsv)          :: stateVectorMeanTrl4D
+  type(struct_gsv)          :: stateVector4D
+  type(struct_gsv)          :: stateVector4Dmod
+  type(struct_gsv)          :: stateVectorWithZandP4D
+  type(struct_gsv)          :: stateVectorHeightSfc
+  type(struct_columnData)   :: column
+
+  type(struct_eob), target  :: ensObs, ensObs_mpiglobal
+  type(struct_eob), pointer :: ensObsGain, ensObsGain_mpiglobal
 
   type(struct_vco), pointer :: vco_ens => null()
   type(struct_hco), pointer :: hco_ens => null()
 
-  integer              :: fclos, fnom, fstopc, ierr
-  integer              :: memberIndex, numStep, numBody
-  integer              :: nulnam, dateStamp
-  integer              :: get_max_rss
+  integer :: get_max_rss, fclos, fnom, fstopc, ierr
+  integer :: memberIndex, nulnam, dateStamp
+  integer :: nEnsGain, eigenVectorIndex, memberIndexInEnsObs
+  integer, allocatable :: dateStampList(:)
+
+  logical  :: useModulatedEns
 
   character(len=256)  :: ensFileName
   character(len=9)    :: obsColumnMode
@@ -57,12 +68,16 @@ program midas_ensembleH
   character(len=48)   :: midasMode
   character(len=10)   :: obsFileType
 
-  logical :: beSilent, dealloc
-
   ! namelist variables
   character(len=256) :: ensPathName
+  character(len=20)  :: obsTimeInterpType ! type of time interpolation to obs time
   integer  :: nEns
-  NAMELIST /NAMENSEMBLEH/nEns, ensPathName
+  integer  :: numRetainedEigen ! number of retained eigenValues/Vectors of vertical localization matrix
+                               !   used only when generating modulated ensembles.
+  real(8)  :: vLocalize        ! vertical localization radius (units: ln(Pressure in Pa) or meters)
+                               !   used only when generating modulated ensembles.
+  NAMELIST /NAMENSEMBLEH/nEns, ensPathName, obsTimeInterpType, numRetainedEigen, &
+                         vLocalize
 
   midasMode = 'analysis'
   obsColumnMode = 'ENKFMIDAS'
@@ -86,29 +101,43 @@ program midas_ensembleH
   call ram_setup
 
   ! Setting default namelist variable values
-  nEns             = 10
-  ensPathName      = 'ensemble'
+  nEns                  = 10
+  ensPathName           = 'ensemble'
+  obsTimeInterpType     = 'LINEAR'
+  numRetainedEigen      = 0
+  vLocalize             = -1.0D0
 
   ! Read the namelist
   nulnam = 0
   ierr = fnom(nulnam, './flnml', 'FTN+SEQ+R/O', 0)
   read(nulnam, nml=namensembleh, iostat=ierr)
-  if ( ierr /= 0) call utl_abort('midas-ensembleH: Error reading namelist')
-  if ( mmpi_myid == 0 ) write(*,nml=namensembleh)
+  if (ierr /= 0) call utl_abort('midas-ensembleH: Error reading namelist')
+  if (mmpi_myid == 0) write(*,nml=namensembleh)
   ierr = fclos(nulnam)
+  
+  if (numRetainedEigen < 0) call utl_abort('midas-ensembleH: numRetainedEigen should be ' // &
+                                             'equal or greater than zero')
+
+  useModulatedEns = (numRetainedEigen > 0)
+  if (useModulatedEns .and. vLocalize <= 0) then
+    call utl_abort('midas-ensembleH: vLocalize should be greater than zero for modulated ens')
+  end if
 
   ! Read the observations
-  call obsf_setup( dateStamp, midasMode, obsFileType_opt = obsFileType )
-  if ( obsFileType /= 'BURP' .and. obsFileType /= 'SQLITE' ) then
+  call obsf_setup(dateStamp, midasMode, obsFileType_opt = obsFileType)
+  if (obsFileType /= 'BURP' .and. obsFileType /= 'SQLITE') then
     call utl_abort('midas-ensembleH: only BURP and SQLITE are valid obs file formats')
   end if
 
   ! Use the first ensemble member to initialize datestamp and grid
-  call fln_ensFileName( ensFileName, ensPathName, memberIndex_opt=1 )
+  call fln_ensFileName(ensFileName, ensPathName, memberIndex_opt=1)
 
   ! Setup timeCoord module, get datestamp from ensemble member
-  call tim_setup( fileNameForDate_opt = ensFileName )
-  numStep = tim_nstepobs
+  call tim_setup(fileNameForDate_opt = ensFileName)
+  allocate(dateStampList(tim_nstepobs))
+  call tim_getstamplist(dateStampList,tim_nstepobs,tim_getDatestamp())
+  
+  write(*,*) 'midas-ensembleH: dateStamp = ',dateStamp
 
   !- Initialize variables of the model states
   call gsv_setup
@@ -116,13 +145,13 @@ program midas_ensembleH
   !- Initialize the Ensemble grid
   if (mmpi_myid == 0) write(*,*) ''
   if (mmpi_myid == 0) write(*,*) 'midas-ensembleH: Set hco and vco parameters for ensemble grid'
-  call hco_SetupFromFile( hco_ens, ensFileName, ' ', 'ENSFILEGRID')
-  call vco_setupFromFile( vco_ens, ensFileName )
+  call hco_SetupFromFile(hco_ens, ensFileName, ' ', 'ENSFILEGRID')
+  call vco_setupFromFile(vco_ens, ensFileName)
 
   write(*,*) 'Memory Used: ', get_max_rss()/1024, 'Mb'
 
   ! read in the observations
-  call inn_setupObs( obsSpaceData, hco_ens, obsColumnMode, obsMpiStrategy, midasMode )
+  call inn_setupObs(obsSpaceData, hco_ens, obsColumnMode, obsMpiStrategy, midasMode)
 
   ! Initialize obs error covariances
   call oer_setObsErrors(obsSpaceData, midasMode)
@@ -135,60 +164,136 @@ program midas_ensembleH
   ! Initialize the observation error covariances
   call oer_setObsErrors(obsSpaceData, midasMode) ! IN
 
-  ! Allocate statevector to store an ensemble member (keep distribution as members on native grid)
-  call gsv_allocate( stateVector, numStep, hco_ens, vco_ens, dateStamp_opt=tim_getDateStamp(),  &
-                     mpi_local_opt=.true., mpi_distribution_opt='VarsLevs', &
-                     dataKind_opt=4, allocHeightSfc_opt=.true. )
-
-  call gsv_allocate( statevector_tiles, numStep, hco_ens, vco_ens, dateStamp_opt=tim_getDateStamp(), &
-                     mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
-                     dataKind_opt=4, allocHeightSfc_opt=.true. )
-
   ! Allocate and initialize eob object for storing HX values
-  numBody = obs_numBody(obsSpaceData)
-  call eob_allocate(ensObs, nEns, numBody, obsSpaceData)
+  call eob_allocate(ensObs, nEns, obs_numBody(obsSpaceData), obsSpaceData)
   call eob_zero(ensObs)
-
+  if (useModulatedEns) then
+    nEnsGain = nEns * numRetainedEigen
+    allocate(ensObsGain)
+    call eob_allocate(ensObsGain, nEnsGain, obs_numBody(obsSpaceData), obsSpaceData)
+    call eob_zero(ensObsGain)
+  else
+    ensObsGain => ensObs
+  end if
+  
   ! Set lat, lon, obs values in ensObs
   call eob_setLatLonObs(ensObs)
+  if (useModulatedEns) call eob_setLatLonObs(ensObsGain)
+
+
+  ! Read the sfc height from ensemble member 1
+  call gsv_allocate(stateVectorHeightSfc, 1, hco_ens, vco_ens, dateStamp_opt=tim_getDateStamp(),  &
+                    mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
+                    dataKind_opt=4, allocHeightSfc_opt=.true., varNames_opt=(/'P0','TT'/))
+  call gio_readFromFile(stateVectorHeightSfc, ensFileName, ' ', ' ',  &
+                        containsFullField_opt=.true., readHeightSfc_opt=.true.)
+
+  ! Allocate statevector related to ensemble mean
+  call gsv_allocate(stateVectorMeanTrl4D, tim_nstepobs, hco_ens, vco_ens, &
+                    dateStamp_opt=tim_getDateStamp(),  &
+                    mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
+                    dataKind_opt=4, allocHeightSfc_opt=.true., &
+                    allocHeight_opt=.false., allocPressure_opt=.false.)
+  call gsv_zero(stateVectorMeanTrl4D)
+  call gsv_allocate(stateVector4D, tim_nstepobs, hco_ens, vco_ens, &
+                    dateStamp_opt=tim_getDateStamp(),  &
+                    mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
+                    dataKind_opt=4, allocHeightSfc_opt=.true., &
+                    allocHeight_opt=.false., allocPressure_opt=.false.)
+  call gsv_zero(stateVector4D)
+  if (useModulatedEns) then
+    ! same as stateVector4D
+    call gsv_allocate(stateVector4Dmod, tim_nstepobs, hco_ens, vco_ens, &
+                      dateStamp_opt=tim_getDateStamp(),  &
+                      mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
+                      dataKind_opt=4, allocHeightSfc_opt=.true., &
+                      allocHeight_opt=.false., allocPressure_opt=.false.)
+    call gsv_zero(stateVector4Dmod)
+  end if
+  
+  ! Allocate statevector for storing state with heights and pressures allocated (for s2c_nl)
+  call gsv_allocate(stateVectorWithZandP4D, tim_nstepobs, hco_ens, vco_ens, &
+                    dateStamp_opt=tim_getDateStamp(),  &
+                    mpi_local_opt=.true., mpi_distribution_opt='Tiles', &
+                    dataKind_opt=4, allocHeightSfc_opt=.true.)
+  call gsv_zero(stateVectorWithZandP4D)
+
+  ! Allocate ensembles, read the Trl ensemble
+  call utl_tmg_start(2,'--ReadEnsemble')
+  call ens_allocate(ensembleTrl4D, nEns, tim_nstepobs, hco_ens, vco_ens, dateStampList)
+  call ens_readEnsemble(ensembleTrl4D, ensPathName, biPeriodic=.false.)
+  call utl_tmg_stop(2)
+
+  ! Compute ensemble mean and copy to meanTrl stateVectors
+  call ens_computeMean(ensembleTrl4D)
+  call ens_copyEnsMean(ensembleTrl4D, stateVectorMeanTrl4D)
 
   do memberIndex = 1, nEns
-    write(*,*) ''
-    write(*,*) 'midas-ensembleH: read member ', memberIndex
-    call utl_tmg_start(2,'--ReadEnsemble')
-    call fln_ensFileName( ensFileName, ensPathName, memberIndex_opt=memberIndex, copyToRamDisk_opt=.false.  )
-    call gio_readFile( stateVector, ensFileName, ' ', ' ', containsFullField=.true., &
-                       readHeightSfc_opt=.true. )
-    call gio_fileUnitsToStateUnits( stateVector, containsFullField=.true. )
-
-    call gsv_transposeVarsLevsToTiles(statevector, statevector_tiles)
-    call utl_tmg_stop(2)
 
     write(*,*) ''
-    write(*,*) 'midas-ensembleH: call s2c_nl for member ', memberIndex
-    write(*,*) ''
-    dealloc = .false.
-    if ( memberIndex == nEns ) dealloc = .true.
-    call s2c_nl( stateVector_tiles, obsSpaceData, column, hco_ens, timeInterpType='LINEAR', dealloc_opt=dealloc )
+    write(*,*) 'midas-letkf: apply nonlinear H to ensemble member ', memberIndex
+    write(*,*) 'Memory Used: ', get_max_rss()/1024, 'Mb'
 
-    write(*,*) ''
-    write(*,*) 'midas-ensembleH: apply nonlinear H to member ', memberIndex
-    write(*,*) ''
-    beSilent = .true.
-    if ( memberIndex == 1 ) beSilent = .false.
-    call inn_computeInnovation(column, obsSpaceData, beSilent_opt=beSilent)
+    ! copy 1 member to a stateVector
+    call ens_copyMember(ensembleTrl4D, stateVector4D, memberIndex)
+    call gsv_copy(stateVector4D, stateVectorWithZandP4D, allowVarMismatch_opt=.true., &
+                  beSilent_opt=.true.)
+    call gsv_copyHeightSfc(stateVectorHeightSfc, stateVectorWithZandP4D)
+
+    ! Compute and set Yb in ensObs
+    call s2c_nl(stateVectorWithZandP4D, obsSpaceData, column, hco_ens, &
+                timeInterpType=obsTimeInterpType, dealloc_opt=.false., &
+                beSilent_opt=.true.)
+
+    ! Compute Y-H(X) in OBS_OMP
+    call inn_computeInnovation(column, obsSpaceData, beSilent_opt=.true.)
 
     ! Copy to ensObs: Y-HX for this member
     call eob_setYb(ensObs, memberIndex)
 
-  end do
+    ! Compute and set Yb in ensObsGain
+    do eigenVectorIndex = 1, numRetainedEigen
+      if (mmpi_myid == 0) write(*,*) 'midas-ensembleH: apply nonlinear H to modulated member ', &
+                                        eigenVectorIndex, '/', numRetainedEigen
 
-  call gsv_deallocate( stateVector_tiles )
-  call gsv_deallocate( stateVector )
+      ! modulate the member with eigenvectors of vertical localization matrix
+      call enkf_getModulatedState(stateVector4D, stateVectorMeanTrl4D, &
+                                  vLocalize, numRetainedEigen, nEns, &
+                                  eigenVectorIndex, stateVector4Dmod, &
+                                  beSilent=.true.)
+
+      call gsv_copy(stateVector4Dmod, stateVectorWithZandP4D, allowVarMismatch_opt=.true., &
+                    beSilent_opt=.true.)
+      call gsv_copyHeightSfc(stateVectorHeightSfc, stateVectorWithZandP4D)
+
+      call s2c_nl(stateVectorWithZandP4D, obsSpaceData, column, hco_ens, &
+                  timeInterpType=obsTimeInterpType, dealloc_opt=.false., &
+                  beSilent_opt=.true.)
+
+      ! Compute Y-H(X) in OBS_OMP
+      call inn_computeInnovation(column, obsSpaceData, filterObsAndInitOer_opt=.false., &
+                                 beSilent_opt=.true.)
+
+      ! Copy to ensObsGain: Y-HX for this member
+      memberIndexInEnsObs = (eigenVectorIndex - 1) * nEns + memberIndex
+      call eob_setYb(ensObsGain, memberIndexInEnsObs)
+    end do ! eigenVectorIndex
+    
+  end do
+  call gsv_deallocate(stateVectorWithZandP4D)
+  if (gsv_isAllocated(stateVector4Dmod)) call gsv_deallocate(stateVector4Dmod)
+  call gsv_deallocate(stateVector4D)
+  call gsv_deallocate(stateVectorMeanTrl4D)
+  call gsv_deallocate(stateVectorHeightSfc)
 
   ! Clean and globally communicate obs-related data, then write to files
   call eob_allGather(ensObs,ensObs_mpiglobal)
-  call eob_writeToFiles(ensObs_mpiglobal)
+  call eob_writeToFiles(ensObs_mpiglobal, outputFilenamePrefix='eob_HX', writeObsInfo=.true.)
+  if (useModulatedEns) then
+    allocate(ensObsGain_mpiglobal)
+    call eob_allGather(ensObsGain, ensObsGain_mpiglobal)
+    call eob_writeToFiles(ensObsGain_mpiglobal, outputFilenamePrefix='eobGain_HX', writeObsInfo=.false.)
+  end if
 
   !
   !- MPI, tmg finalize
@@ -204,8 +309,8 @@ program midas_ensembleH
   !
   !- 7.  Ending
   !
-  if ( mmpi_myid == 0 ) write(*,*) ' --------------------------------'
-  if ( mmpi_myid == 0 ) write(*,*) ' MIDAS-ENSEMBLEH ENDS'
-  if ( mmpi_myid == 0 ) write(*,*) ' --------------------------------'
+  if (mmpi_myid == 0) write(*,*) ' --------------------------------'
+  if (mmpi_myid == 0) write(*,*) ' MIDAS-ENSEMBLEH ENDS'
+  if (mmpi_myid == 0) write(*,*) ' --------------------------------'
 
 end program midas_ensembleH